# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudentHub Africa API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (students for marketplace)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, company, individual]
        - name: skill
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, company, individual]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        averageRating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, company, individual]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        portfolioUrl:
          type: ["string", "null"]
        cvUrl:
          type: ["string", "null"]
        averageRating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        completedMissions:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserList:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        university:
          type: string
        skills:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
        portfolioUrl:
          type: string
        cvUrl:
          type: string

    # SERVICE
...